Lecture 4
Learned MVVM(enum and special type of enum is optional).
Views are declarative and reactive with the help of observedObject.
Model is able to store the states.Therefore, we have removed the @states from the view.
View is completely reactive and stateless.
viewModel is a keeper(@Published etc) & Interpreter as well (choose method )
Structs able to detect the changes.